add -a flag

+++ WM/src/printeble.c
@@ -27,8 +27,11 @@ void	print_last_alive(void)
 {
 	ft_printf("Player %d \"%s\" won!\n", st.last_alive + 1,
 		st.champ[st.last_alive].name);
-	kill_sounds();
-	system("rm temp");
+	if (st.flag_visual)
+	{
+		kill_sounds();
+		system("rm temp");
+	}
 }
 
 void	print_field(int i)
@@ -69,7 +72,6 @@ void	print_field(int i)
 		ft_printf("\n");
 		a += i;
 	}
-	system("leaks corewar");
 }
 
 void	logs(int i, int d)
